book-service/requirements.txt
django
djangorestframework
